Only use Load from ws history if final command
Loading a Workspace can also load the Workspace's history. When using
this history to restore a Workspace, much of this history may be
irrelavant or reference other Workspaces that do not exist on the
current machine.

If the final command in the Workspace's history is a Load command, only
this load command is used to restore the Workspace.