Ensure variable names are valid in workspace history
Add clean_variable_name function to ensure a string is a valid Python
identifier before assigning to it via operations in the workspace
history. Previously the workspace name was used the variable name, but a
workspace name is not always a valid variable name.

Hit test_plotscriptgeneratorutils.py with YAPF.