Fix failing zorder test in earlier MPL version
A test that checked if the "barsabove" argument was correct within the
plot commands was failing on earlier versions of matplotlib. This was
because lines in ErrorbarContainers were given the same zorder in
earlier version.

Since the error bars are plotted after the connecting line, use the >=
comparator on the zorders.